+++ envolved/__init__.py
@@ -1,6 +1,7 @@
 from envolved._version import __version__
-from envolved.schema import Schema
 from envolved.envvar import EnvVar
 from envolved.exceptions import MissingEnvError
+from envolved.describe import describe_env_vars
+from envolved.schema import Schema
 
-__all__ = ['__version__', 'Schema', 'EnvVar', 'MissingEnvError']
+__all__ = ['__version__', 'Schema', 'EnvVar', 'MissingEnvError', 'describe_env_vars']