+++ envolved/basevar.py
@@ -1,5 +1,6 @@
 from abc import abstractmethod
-from typing import Generic, TypeVar, Union, List, Callable, Any, Type
+from textwrap import TextWrapper
+from typing import Generic, TypeVar, Union, List, Callable, Any, Type, Optional
 from weakref import ref
 
 from envolved.envparser import env_parser, CaseInsensitiveAmbiguity
@@ -110,6 +111,10 @@ class EnvironmentVariable(BaseVar[T], Generic[T]):
         self._validators.append(callback)
         return func
 
+    @abstractmethod
+    def description(self, parent_wrapper: TextWrapper) -> List[str]:
+        pass
+
 
 class SingleKeyEnvVar(EnvironmentVariable[T], Generic[T]):
     """
@@ -117,17 +122,20 @@ class SingleKeyEnvVar(EnvironmentVariable[T], Generic[T]):
     """
 
     def __init__(self, key: str, default: T, *,
-                 case_sensitive: bool = False, type: Union[Type[T], Parser[T]] = str):
+                 case_sensitive: bool = False, type: Union[Type[T], Parser[T]] = str,
+                 description: Optional[str] = None):
         """
-        :param key: The external name of the environment variable
-        :param default: passed to EnvironmentVariable.__init__
-        :param case_sensitive: Whether the name is case sensitive or not
-        :param type: The internal conversion type or parser from the string value of the var to the output type
+        :param key: The external name of the environment variable.
+        :param default: passed to EnvironmentVariable.__init__.
+        :param case_sensitive: Whether the name is case sensitive or not.
+        :param type: The internal conversion type or parser from the string value of the var to the output type.
+        :param description: Description of the variable.
         """
         super().__init__(default)
         self.key = key
         self.converter = parser(type)
         self.case_sensitive = case_sensitive
+        self._description = description
 
     def _get(self) -> T:
         try:
@@ -140,3 +148,13 @@ class SingleKeyEnvVar(EnvironmentVariable[T], Generic[T]):
         raw_value = raw_value.strip()
 
         return self.converter(raw_value)
+
+    def description(self, parent_wrapper: TextWrapper) -> List[str]:
+        key = self.key
+        if not self.case_sensitive:
+            key = key.upper()
+
+        if not self._description:
+            return [key]
+        desc = ' '.join(self._description.strip().split())
+        return parent_wrapper.wrap(key+': '+desc)